extract intermediate terraform class
